docs: Wave 115 deep-debt evolution — root docs update, deprecated cleanup, cargo clean
README: 471 tests, depot.integrity command, gate/{mesh,nucleus} modules, Wave 115
hardening status. GLACIAL_SHIFT_TRACKER: Wave 115 update header. VPS_STATE,
IRONGATE_VERIFICATION, RUNBOOKS: bumped to Wave 115.

wateringHole handoff: cellMembrane code evolution table (14 items shipped),
bootstrap/depot fixes checked off, identity.git detection marked done.

Removed 3 unused DEPRECATED fallback constants (FALLBACK_CRYPTO_SIGNER,
FALLBACK_MESH_RELAY, FALLBACK_CONTENT_SERVING) — zero consumers since Wave 110.

cargo clean: reclaimed 1.8 GB build artifacts.

+++ crates/membrane-shadow/src/forgejo.rs
@@ -15,7 +15,7 @@ use serde::{Deserialize, Serialize};
 const API_TIMEOUT_WRITE: std::time::Duration = std::time::Duration::from_secs(30);
 const API_TIMEOUT_READ: std::time::Duration = std::time::Duration::from_secs(15);
 const API_TIMEOUT_FAST: std::time::Duration = std::time::Duration::from_secs(5);
-const PAGE_SIZE: u32 = 50;
+const PAGE_SIZE: usize = 50;
 
 // ── Types ───────────────────────────────────────────────────────────
 
@@ -195,7 +195,7 @@ pub async fn repo_list(config: &ShadowConfig, org: &str) -> Result<Vec<RepoInfo>
         let batch: Vec<RepoInfo> = resp.json().await?;
         let count = batch.len();
         all_repos.extend(batch);
-        if count < PAGE_SIZE as usize {
+        if count < PAGE_SIZE {
             break;
         }
         page += 1;
